Prepare for remote decryption. Rename DecryptingMediator to Decryption. Add tallyResult() and decryptionResult() to ElectionRecord. Tweak the comments. Move RandomBallotProvider to electionguard input, so can be used in remote testing in Java.
